+++ src/lerobot/rl/crop_dataset_roi.py
