+++ src/lerobot/rl/eval_policy.py
@@ -25,11 +25,12 @@ from lerobot.robots import (  # noqa: F401
     make_robot_from_config,
     so100_follower,
 )
-from lerobot.scripts.rl.gym_manipulator import make_robot_env
 from lerobot.teleoperators import (
     gamepad,  # noqa: F401
     so101_leader,  # noqa: F401
 )
+
+from .gym_manipulator import make_robot_env
 
 logging.basicConfig(level=logging.INFO)
